# server/docs/swagger.yaml
basePath: /
definitions:
...
  api.Error:
    type: object
...
  model.Category:
    properties:
      id:
        example: 999
        type: integer
      insert_time:
        type: string
      last_update_time:
        type: string
      name:
        type: string
    required:
    - id
    - name
    type: object
...
  model.Collection:
    properties:
      categories:
        items:
          type: integer
        type: array
      citation_template:
        type: string
      detailPrivacy:
        type: string
      fields:
        items:
          $ref: '#/definitions/model.CollectionField'
        type: array
      genderHeader:
        type: string
      householdNumberHeader:
        type: string
      householdRelationshipHeader:
        type: string
      id:
        example: 999
        type: integer
      imagePathHeader:
        type: string
      imagePrivacy:
        type: string
      indexPrivacy:
        type: string
      insert_time:
        type: string
      last_update_time:
        type: string
      location:
        type: string
      mappings:
        items:
          $ref: '#/definitions/model.CollectionMapping'
        type: array
      name:
        type: string
    required:
    - categories
    - id
    - name
    type: object
...
host: api.ourroots.org
info:
  contact:
    email: jim@anconafamily.com
    name: Jim Ancona
    url: https://github.com/jancona
  description: This is the OurRoots API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: OurRoots API
  version: 0.1.0
paths:
...
  /categories/{id}:
    delete:
      operationId: deleteCategory
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: OK
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: deletes a Category
      tags:
      - categories
    get:
      operationId: getCategory
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Category'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: gets a Category
      tags:
      - categories
    put:
      consumes:
      - application/json
      operationId: updateCategory
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/model.Category'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Category'
        "415":
          description: Bad Content-Type
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: updates a Category
      tags:
      - categories
  /collections:
    get:
      operationId: getCollections
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Collection'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: returns all collections
      tags:
      - collections
    post:
      consumes:
      - application/json
      operationId: addCollection
      parameters:
      - description: Add Collection
        in: body
        name: collection
        required: true
        schema:
          $ref: '#/definitions/model.CollectionIn'
      produces:
      - application/json
      responses:
        "201":
          description: OK
          schema:
            $ref: '#/definitions/model.Collection'
        "415":
          description: Bad Content-Type
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: adds a new Collection
      tags:
      - collections
  /collections/{id}:
    delete:
      operationId: deleteCollection
      parameters:
      - description: Collection ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: OK
          schema:
            $ref: '#/definitions/model.Collection'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: deletes a Collection
      tags:
      - collections
    get:
      operationId: getCollection
      parameters:
      - description: Collection ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Collection'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: gets a Collection
      tags:
      - collections
    put:
      consumes:
      - application/json
      operationId: updateCollection
      parameters:
      - description: Collection ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Collection
        in: body
        name: collection
        required: true
        schema:
          $ref: '#/definitions/model.Collection'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Collection'
        "415":
          description: Bad Content-Type
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: updates a Collection
      tags:
      - collections
  /content:
    post:
      consumes:
      - application/json
      operationId: postContentRequest
      parameters:
      - description: Create content request
        in: body
        name: contentRequest
        required: true
        schema:
          $ref: '#/definitions/api.ContentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Category'
        "415":
          description: Bad Content-Type
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: returns a URL for uploading content
      tags:
      - content
  /currentuser:
    get:
      operationId: getCurrentUser
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Place'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: returns the current user
      tags:
      - users
  /invitations/{code}:
    get:
      operationId: getInvitation
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.InvitationSocietyName'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: gets an Invitation
      tags:
      - invitations
    post:
      consumes:
      - application/json
      operationId: acceptInvitation
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SocietyUser'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: accept an invitation
      tags:
      - invitations
  /places:
    get:
      operationId: getPlacesByPrefix
      parameters:
      - description: place prefix
        in: query
        name: prefix
        type: string
      - description: maximum number of places to return
        in: query
        name: count
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Place'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: returns places matching prefix
      tags:
      - places
  /posts:
    get:
      operationId: getPosts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Post'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: returns all posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      operationId: addPost
      parameters:
      - description: Add Post
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/model.PostIn'
      produces:
      - application/json
      responses:
        "201":
          description: OK
          schema:
            $ref: '#/definitions/model.Post'
        "415":
          description: Bad Content-Type
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: adds a new Post
      tags:
      - posts
  /posts/{id}:
    delete:
      operationId: deletePost
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: OK
          schema:
            $ref: '#/definitions/model.Post'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: deletes a Post
      tags:
      - posts
    get:
      operationId: getPost
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Post'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: gets a Post
      tags:
      - posts
    put:
      consumes:
      - application/json
      operationId: updatePost
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Post
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/model.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Post'
        "415":
          description: Bad Content-Type
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: updates a Post
      tags:
      - posts
  /posts/{id}/images/{filePath}:
    get:
      operationId: getPostImage
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: Image file path
        in: path
        name: imageFile
        required: true
        type: string
      - description: return the url as json {url, height, width} if true
        in: query
        name: noredirect
        type: boolean
      - description: return thumbnail
        in: query
        name: thumbnail
        type: boolean
      responses:
        "307":
          description: Temporary Redirect
          schema:
            type: header
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: Returns a redirect to an image URL
      tags:
      - posts
  /records:
    get:
      operationId: getRecords
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.RecordsResult'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: returns all records
      tags:
      - records
...
schemes:
- http
- https
securityDefinitions:
  OAuth2Implicit:
    authorizationUrl: https://ourroots.auth0.com/authorize?audience=https%3A%2F%2Fapi.ourroots.org%2Fpreprod
    flow: implicit
    scopes:
      cms: ' Grants read and write access to the CMS'
      email: ' Grants access to OIDC email attributes'
      openid: ' Indicates that the application intends to use OIDC to verify the user''s
        identity'
      profile: ' Grants access to OIDC user profile attributes'
    type: oauth2
swagger: "2.0"